한 라인의 값을 담당하는 Line 클래스 구현
+ 사다리 데이터를 String 이 아닌 Line 객체로 받도록 리팩토링
(List<String> -> List<Line>)
